Update documentation & refactor the schema & fields into a single file.
